src/gemini/requirements.txt
google-generativeai>=0.3.0
pillow>=10.0.0
python-dotenv>=1.0.0 